removed unnecessary comments
