+++ pkg/core/proxy/integrations/http/decode.go
@@ -13,8 +13,6 @@ import (
 	"net"
 	"net/http"
 	"strconv"
-	"strings"
-	"time"
 
 	"go.keploy.io/server/v2/pkg"
 	"go.keploy.io/server/v2/pkg/core/proxy/integrations"
@@ -199,161 +197,5 @@ func (h *HTTP) decodeHTTP(ctx context.Context, logger *zap.Logger, reqBuf []byte
 			return nil
 		}
 		return err
-	}
-}
-
-func (h *HTTP) HandleChunkedRequests(ctx context.Context, logger *zap.Logger, finalReq *[]byte, clientConn, destConn net.Conn) error {
-
-	if hasCompleteHeaders(*finalReq) {
-		logger.Debug("this request has complete headers in the first chunk itself.")
-	}
-
-	for !hasCompleteHeaders(*finalReq) {
-		logger.Debug("couldn't get complete headers in first chunk so reading more chunks")
-		reqHeader, err := pUtil.ReadBytes(ctx, logger, clientConn)
-		if err != nil {
-			utils.LogError(logger, nil, "failed to read the request message from the client")
-			return err
-		}
-		// destConn is nil in case of test mode
-		if destConn != nil {
-			_, err = destConn.Write(reqHeader)
-			if err != nil {
-				if ctx.Err() != nil {
-					return ctx.Err()
-				}
-				utils.LogError(logger, nil, "failed to write request message to the destination server")
-				return err
-			}
-		}
-
-		*finalReq = append(*finalReq, reqHeader...)
-	}
-
-	lines := strings.Split(string(*finalReq), "\n")
-	var contentLengthHeader string
-	var transferEncodingHeader string
-	for _, line := range lines {
-		if strings.HasPrefix(line, "Content-Length:") {
-			contentLengthHeader = strings.TrimSpace(strings.TrimPrefix(line, "Content-Length:"))
-			break
-		} else if strings.HasPrefix(line, "Transfer-Encoding:") {
-			transferEncodingHeader = strings.TrimSpace(strings.TrimPrefix(line, "Transfer-Encoding:"))
-			break
-		}
-	}
-
-	//Handle chunked requests
-	if contentLengthHeader != "" {
-		contentLength, err := strconv.Atoi(contentLengthHeader)
-		if err != nil {
-			utils.LogError(logger, err, "failed to get the content-length header")
-			return fmt.Errorf("failed to handle chunked request")
-		}
-		//Get the length of the body in the request.
-		bodyLength := len(*finalReq) - strings.Index(string(*finalReq), "\r\n\r\n") - 4
-		contentLength -= bodyLength
-		if contentLength > 0 {
-			err := h.contentLengthRequest(ctx, logger, finalReq, clientConn, destConn, contentLength)
-			if err != nil {
-				return err
-			}
-		}
-	} else if transferEncodingHeader != "" {
-		// check if the initial request is the complete request.
-		if strings.HasSuffix(string(*finalReq), "0\r\n\r\n") {
-			return nil
-		}
-		if transferEncodingHeader == "chunked" {
-			err := h.chunkedRequest(ctx, logger, finalReq, clientConn, destConn, transferEncodingHeader)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// Handled chunked requests when content-length is given.
-func (h *HTTP) contentLengthRequest(ctx context.Context, logger *zap.Logger, finalReq *[]byte, clientConn, destConn net.Conn, contentLength int) error {
-	for contentLength > 0 {
-		err := clientConn.SetReadDeadline(time.Now().Add(5 * time.Second))
-		if err != nil {
-			utils.LogError(logger, err, "failed to set the read deadline for the client conn")
-			return err
-		}
-		requestChunked, err := pUtil.ReadBytes(ctx, logger, clientConn)
-		if err != nil {
-			if err == io.EOF {
-				utils.LogError(logger, nil, "conn closed by the user client")
-				return err
-			} else if netErr, ok := err.(net.Error); ok && netErr.Timeout() {
-				logger.Info("Stopped getting data from the conn", zap.Error(err))
-				break
-			}
-			utils.LogError(logger, nil, "failed to read the response message from the destination server")
-			return err
-		}
-		logger.Debug("This is a chunk of request[content-length]: " + string(requestChunked))
-		*finalReq = append(*finalReq, requestChunked...)
-		contentLength -= len(requestChunked)
-
-		// destConn is nil in case of test mode.
-		if destConn != nil {
-			_, err = destConn.Write(requestChunked)
-			if err != nil {
-				if ctx.Err() != nil {
-					return ctx.Err()
-				}
-				utils.LogError(logger, nil, "failed to write request message to the destination server")
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// Handled chunked requests when transfer-encoding is given.
-func (h *HTTP) chunkedRequest(ctx context.Context, logger *zap.Logger, finalReq *[]byte, clientConn, destConn net.Conn, _ string) error {
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-			//TODO: we have to implement a way to read the buffer chunk wise according to the chunk size (chunk size comes in hexadecimal)
-			// because it can happen that some chunks come after 5 seconds.
-			err := clientConn.SetReadDeadline(time.Now().Add(5 * time.Second))
-			if err != nil {
-				utils.LogError(logger, err, "failed to set the read deadline for the client conn")
-				return err
-			}
-			requestChunked, err := pUtil.ReadBytes(ctx, logger, clientConn)
-			if err != nil {
-				if netErr, ok := err.(net.Error); ok && netErr.Timeout() {
-					break
-				}
-				utils.LogError(logger, nil, "failed to read the response message from the destination server")
-				return err
-			}
-
-			*finalReq = append(*finalReq, requestChunked...)
-			// destConn is nil in case of test mode.
-			if destConn != nil {
-				_, err = destConn.Write(requestChunked)
-				if err != nil {
-					if ctx.Err() != nil {
-						return ctx.Err()
-					}
-					utils.LogError(logger, nil, "failed to write request message to the destination server")
-					return err
-				}
-			}
-
-			//check if the initial request is completed
-			if strings.HasSuffix(string(requestChunked), "0\r\n\r\n") {
-				return nil
-			}
-		}
 	}
 }
