+++ pkg/core/proxy/integrations/http/encode.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"io"
 	"net"
-	"strconv"
 	"strings"
 	"time"
 
@@ -260,179 +259,4 @@ func (h *HTTP) encodeHTTP(ctx context.Context, logger *zap.Logger, reqBuf []byte
 		}
 		return err
 	}
-}
-
-func (h *HTTP) handleChunkedResponses(ctx context.Context, logger *zap.Logger, finalResp *[]byte, clientConn, destConn net.Conn, resp []byte) error {
-
-	if hasCompleteHeaders(*finalResp) {
-		logger.Debug("this response has complete headers in the first chunk itself.")
-	}
-
-	for !hasCompleteHeaders(resp) {
-		logger.Debug("couldn't get complete headers in first chunk so reading more chunks")
-		respHeader, err := pUtil.ReadBytes(ctx, logger, destConn)
-		if err != nil {
-			if err == io.EOF {
-				logger.Debug("received EOF from the server")
-				// if there is any buffer left before EOF, we must send it to the client and save this as mock
-				if len(respHeader) != 0 {
-					// write the response message to the user client
-					_, err = clientConn.Write(resp)
-					if err != nil {
-						if ctx.Err() != nil {
-							return ctx.Err()
-						}
-						utils.LogError(logger, nil, "failed to write response message to the user client")
-						return err
-					}
-					*finalResp = append(*finalResp, respHeader...)
-				}
-				return err
-			}
-			utils.LogError(logger, nil, "failed to read the response message from the destination server")
-			return err
-		}
-		// write the response message to the user client
-		_, err = clientConn.Write(respHeader)
-		if err != nil {
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			utils.LogError(logger, nil, "failed to write response message to the user client")
-			return err
-		}
-
-		*finalResp = append(*finalResp, respHeader...)
-		resp = append(resp, respHeader...)
-	}
-
-	//Getting the content-length or the transfer-encoding header
-	var contentLengthHeader, transferEncodingHeader string
-	lines := strings.Split(string(resp), "\n")
-	for _, line := range lines {
-		if strings.HasPrefix(line, "Content-Length:") {
-			contentLengthHeader = strings.TrimSpace(strings.TrimPrefix(line, "Content-Length:"))
-			break
-		} else if strings.HasPrefix(line, "Transfer-Encoding:") {
-			transferEncodingHeader = strings.TrimSpace(strings.TrimPrefix(line, "Transfer-Encoding:"))
-			break
-		}
-	}
-	//Handle chunked responses
-	if contentLengthHeader != "" {
-		contentLength, err := strconv.Atoi(contentLengthHeader)
-		if err != nil {
-			utils.LogError(logger, err, "failed to get the content-length header")
-			return fmt.Errorf("failed to handle chunked response")
-		}
-		bodyLength := len(resp) - strings.Index(string(resp), "\r\n\r\n") - 4
-		contentLength -= bodyLength
-		if contentLength > 0 {
-			err := h.contentLengthResponse(ctx, logger, finalResp, clientConn, destConn, contentLength)
-			if err != nil {
-				return err
-			}
-		}
-	} else if transferEncodingHeader != "" {
-		//check if the initial response is the complete response.
-		if strings.HasSuffix(string(*finalResp), "0\r\n\r\n") {
-			return nil
-		}
-		if transferEncodingHeader == "chunked" {
-			err := h.chunkedResponse(ctx, logger, finalResp, clientConn, destConn)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// Handled chunked responses when transfer-encoding is given.
-func (h *HTTP) chunkedResponse(ctx context.Context, logger *zap.Logger, finalResp *[]byte, clientConn, destConn net.Conn) error {
-	isEOF := false
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-			resp, err := pUtil.ReadBytes(ctx, logger, destConn)
-			if err != nil {
-				if err != io.EOF {
-					utils.LogError(logger, err, "failed to read the response message from the destination server")
-					return err
-				}
-				isEOF = true
-				logger.Debug("received EOF", zap.Error(err))
-				if len(resp) == 0 {
-					logger.Debug("exiting loop as response is complete")
-					break
-				}
-			}
-
-			*finalResp = append(*finalResp, resp...)
-			// write the response message to the user client
-			_, err = clientConn.Write(resp)
-			if err != nil {
-				if ctx.Err() != nil {
-					return ctx.Err()
-				}
-				utils.LogError(logger, nil, "failed to write response message to the user client")
-				return err
-			}
-
-			//In some cases need to write the response to the client
-			// where there is some response before getting the true EOF
-			if isEOF {
-				break
-			}
-
-			if string(resp) == "0\r\n\r\n" {
-				return nil
-			}
-		}
-	}
-}
-
-
-// Handled chunked responses when content-length is given.
-func (h *HTTP) contentLengthResponse(ctx context.Context, logger *zap.Logger, finalResp *[]byte, clientConn, destConn net.Conn, contentLength int) error {
-	isEOF := false
-	for contentLength > 0 {
-		resp, err := pUtil.ReadBytes(ctx, logger, destConn)
-		if err != nil {
-			if err == io.EOF {
-				isEOF = true
-				logger.Debug("received EOF, conn closed by the destination server")
-				if len(resp) == 0 {
-					break
-				}
-			} else if netErr, ok := err.(net.Error); ok && netErr.Timeout() {
-				logger.Info("Stopped getting data from the conn", zap.Error(err))
-				break
-			} else {
-				utils.LogError(logger, nil, "failed to read the response message from the destination server")
-				return err
-			}
-		}
-
-		logger.Debug("This is a chunk of response[content-length]: " + string(resp))
-		*finalResp = append(*finalResp, resp...)
-		contentLength -= len(resp)
-
-		// write the response message to the user client
-		_, err = clientConn.Write(resp)
-		if err != nil {
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			utils.LogError(logger, nil, "failed to write response message to the user client")
-			return err
-		}
-
-		if isEOF {
-			break
-		}
-	}
-	return nil
 }
